+++ StarkBank/StarkBankTests/DepositLog.cs
@@ -1,0 +1,35 @@
+﻿using Xunit;
+using System;
+using StarkBank;
+using System.Linq;
+using System.Collections.Generic;
+
+
+namespace StarkBankTests
+{
+    public class DepositLogTest
+    {
+        public readonly User user = TestUser.SetDefault();
+
+        [Fact]
+        public void QueryAndGet()
+        {
+            List<Deposit.Log> logs = Deposit.Log.Query(
+                limit: 101,
+                before: DateTime.Now.Date,
+                types: new List<string> { "created" }
+            ).ToList();
+            Assert.Equal(101, logs.Count);
+            Assert.True(logs.First().ID != logs.Last().ID);
+            foreach (Deposit.Log log in logs)
+            {
+                TestUtils.Log(log);
+                Assert.NotNull(log.ID);
+                Assert.Equal("created", log.Type);
+            }
+            Invoice.Log getLog = Invoice.Log.Get(id: logs.First().ID);
+            Assert.Equal(getLog.ID, logs.First().ID);
+            TestUtils.Log(getLog);
+        }
+    }
+}